Chore: Move shareRequest to message instead of schemas
