+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './credentialList';
+export * from './messages';
 export * from './validator';
 export * from './schemas';
