+++ src/schemas/index.ts
@@ -2,5 +2,4 @@ export * as benefit from './benefit';
 export * as finance from './finance';
 export * as identity from './identity';
 export * as livingPlace from './livingPlace';
-export * as shareRequest from './shareRequest';
 export * as work from './work';
